chore: move selectors
