+++ ui/src/alerting/components/AlertHistoryIndex.tsx
@@ -23,7 +23,9 @@ import {
   getInitialHistoryType,
   getInitialState,
 } from 'src/alerting/utils/history'
-import {getCheckIDs, getEndpointIDs, getRuleIDs} from 'src/alerting/selectors'
+import {getCheckIDs} from 'src/checks/selectors'
+import {getEndpointIDs} from 'src/notifications/endpoints/selectors'
+import {getRuleIDs} from 'src/notifications/rules/selectors'
 
 // Types
 import {ResourceIDs} from 'src/checks/reducers'
