+++ ui/src/notifications/rules/selectors/index.ts
@@ -1,0 +1,5 @@
+import {AppState} from 'src/types'
+
+export const getRuleIDs = (state: AppState): {[x: string]: boolean} => {
+  return state.rules.list.reduce((acc, rule) => ({...acc, [rule.id]: true}), {})
+}
